+++ examples/konga_default_password/main.tf
@@ -1,14 +1,19 @@
 module konga {
-  source  = "../../"
+  source = "../../"
 
   db_host = var.db_host
   db_name = var.db_name
   db_pass = var.db_pass
   db_user = var.db_user
 
-  # we can autoconfigure a default connection pass kong_url and if we want also kong_name variables. 
-  kong_url = "http://kong:8001"
-
+  # we can autoconfigure a default connection pass kong_admin_url and if we want also name as keys for kong_endpoint variable.
+  kong_endpoints = [
+    {
+      "kong_admin_url" = "http://kong:8001",
+      "name"           = "kong-admin",
+      "type"           = "default"
+    },
+  ]
 
   # When enable_ingress is true konga is expose whitin an ingress controller, is possible to use annotations also. 
   enable_ingress = true
@@ -33,5 +38,5 @@ variable db_pass {}
 
 output password {
   sensitive = true
-  value = module.konga.password
+  value     = module.konga.password
 }